requirements.txt
boto3>=1.19
Click>=8.0
yaspin==0.14.1
deepmerge
PyYAML
requests
six>=1.15
questionary>=1.10
